app_sicpej/requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.2
djangorestframework==3.16.1
pillow==11.2.1
reportlab==4.4.1
sqlparse==0.5.3
